[ci skip] sad casting moment

+++ core/common/src/main/java/org/screamingsandals/lib/block/BlockPosition.java
@@ -17,9 +17,9 @@ public class BlockPosition {
 
     public static long toLong(int x, int y, int z) {
         if (Server.isVersion(1, 14)) {
-            return ((long) (x & 0x3FFFFFF) << 38) | ((long) (z & 0x3FFFFFF) << 12) | (y & 0xFFF);
+            return (((long) x & 0x3FFFFFF) << 38) | (((long) z & 0x3FFFFFF) << 12) | ((long) y & 0xFFF);
         }
-        return ((long) (x & 0x3FFFFFF) << 38) | ((long) (y & 0xFFF) << 26) | (z & 0x3FFFFFF);
+        return (((long) x & 0x3FFFFFF) << 38) | (((long) y & 0xFFF) << 26) | ((long) z & 0x3FFFFFF);
     }
 
     public long asLong() {
